requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
requests
openai
langdetect